Destroy vue.js instances (components) when the GWT widget is disposed
Refactored middleware and moved component creation to a separate factory.

`SubscribedLabel` shows how a component can lister to global events.
When the component is disposed, the event subscriptions are removed.
We "prove" this by logging messages when receiving the event. Adding,
removing and adding a component proves that there's no leak.